Remove penalty for break/continue
According to the Cognitive Complexity specification, only `break` and
`continue` _to a label_ add a fundamental increment. Simple forms of
these are easier to follow and "often make code much clearer", so
these do not (of themselves) add any increment. (These keywords tend
to appear in conditional constructs, and the added complexity is
already embodied in the structural increment for the conditional.)