Fix overcounting for decorators
The Cognitive Complexity specification makes an exception for
decorators (a function that contains only a nested function and a
return statement). In this case, there is no penalty for the nesting.
To fix overcounting in this case, detect a decorator in
api.get_cognitive_complexity() as this is where the top-level function
is handled.